refactor: move app providers into sub folder in browser lib
